+++ lib/wraft_doc/organisation/enterprise.ex
@@ -17,6 +17,13 @@ defmodule WraftDoc.Enterprise do
     Document
   }
 
+  @default_states [%{"state" => "Draft", "order" => 1}, %{"state" => "Publish", "order" => 2}]
+  @default_controlled_states [
+    %{"state" => "Draft", "order" => 1},
+    %{"state" => "Review", "order" => 2},
+    %{"state" => "Publish", "order" => 3}
+  ]
+
   @doc """
   Get a flow from its UUID.
   """
@@ -34,10 +41,32 @@ defmodule WraftDoc.Enterprise do
   end
 
   @doc """
-  Create a flow.
+  Create a controlled flow flow.
   """
   @spec create_flow(User.t(), map) ::
           %Flow{creator: User.t()} | {:error, Ecto.Changeset.t()}
+
+  def create_flow(%{organisation_id: org_id} = current_user, %{"controlled" => true} = params) do
+    params = params |> Map.merge(%{"organisation_id" => org_id})
+
+    current_user
+    |> build_assoc(:flows)
+    |> Flow.controlled_changeset(params)
+    |> Spur.insert()
+    |> case do
+      {:ok, flow} ->
+        Task.start_link(fn -> create_default_states(current_user, flow, true) end)
+        flow |> Repo.preload(:creator)
+
+      {:error, _} = changeset ->
+        changeset
+    end
+  end
+
+  @doc """
+  Create an uncontrolled flow flow.
+  """
+
   def create_flow(%{organisation_id: org_id} = current_user, params) do
     params = params |> Map.merge(%{"organisation_id" => org_id})
 
@@ -47,6 +76,7 @@ defmodule WraftDoc.Enterprise do
     |> Spur.insert()
     |> case do
       {:ok, flow} ->
+        Task.start_link(fn -> create_default_states(current_user, flow) end)
         flow |> Repo.preload(:creator)
 
       {:error, _} = changeset ->
@@ -76,12 +106,11 @@ defmodule WraftDoc.Enterprise do
   end
 
   @doc """
-  Update a flow.
+  Update a controlled flow
   """
-  @spec update_flow(Flow.t(), User.t(), map) :: Flow.t() | {:error, Ecto.Changeset.t()}
-  def update_flow(flow, %User{id: id}, params) do
+  def update_flow(flow, %User{id: id}, %{"controlled" => true} = params) do
     flow
-    |> Flow.changeset(params)
+    |> Flow.update_controlled_changeset(params)
     |> Spur.update(%{actor: "#{id}"})
     |> case do
       {:ok, flow} ->
@@ -93,7 +122,24 @@ defmodule WraftDoc.Enterprise do
   end
 
   @doc """
-  Delete a flow.
+  Update a uncontrolled flow.
+  """
+  @spec update_flow(Flow.t(), User.t(), map) :: Flow.t() | {:error, Ecto.Changeset.t()}
+  def update_flow(flow, %User{id: id}, params) do
+    flow
+    |> Flow.update_changeset(params)
+    |> Spur.update(%{actor: "#{id}"})
+    |> case do
+      {:ok, flow} ->
+        flow |> Repo.preload(:creator)
+
+      {:error, _} = changeset ->
+        changeset
+    end
+  end
+
+  @doc """
+  Delete a  flow.
   """
   @spec delete_flow(Flow.t(), User.t()) :: {:ok, Flow.t()} | {:error, Ecto.Changeset.t()}
   def delete_flow(flow, %User{id: id}) do
@@ -108,12 +154,37 @@ defmodule WraftDoc.Enterprise do
   end
 
   @doc """
+  Create default states for a controlled fow
+  """
+
+  @spec create_default_states(User.t(), Flow.t(), boolean()) :: list
+  def create_default_states(current_user, flow, true) do
+    Enum.map(@default_controlled_states, fn x -> create_state(current_user, flow, x) end)
+  end
+
+  @doc """
+  Create default states for an uncontrolled flow
+  """
+
+  def create_default_states(current_user, flow) do
+    Enum.map(@default_states, fn x -> create_state(current_user, flow, x) end)
+  end
+
+  @doc """
   Create a state under a flow.
   """
-  @spec create_state(User.t(), Flow.t(), map) :: {:ok, State.t()} | {:error, Ecto.Changeset.t()}
+  @spec create_state(User.t(), Flow.t(), map) :: State.t() | {:error, Ecto.Changeset.t()}
   def create_state(%User{organisation_id: org_id} = current_user, flow, params) do
     params = params |> Map.merge(%{"organisation_id" => org_id})
-    current_user |> build_assoc(:states, flow: flow) |> State.changeset(params) |> Spur.insert()
+
+    current_user
+    |> build_assoc(:states, flow: flow)
+    |> State.changeset(params)
+    |> Spur.insert()
+    |> case do
+      {:ok, state} -> state
+      {:error, _} = changeset -> changeset
+    end
   end
 
   @doc """
@@ -411,9 +482,26 @@ defmodule WraftDoc.Enterprise do
   end
 
   @doc """
+  Check the user and approver is same while approving the content
+  """
+  def same_user?(current_user_uuid, approver_uuid) when current_user_uuid != approver_uuid,
+    do: :invalid_user
+
+  def same_user?(current_user_uuid, approver_uuid) when current_user_uuid === approver_uuid,
+    do: true
+
+  @doc """
+  Check the prestate of the approval system and state of instance are same
+  """
+  def same_state?(prestate_id, state_id) when prestate_id != state_id,
+    do: :unprocessible_state
+
+  def same_state?(prestate_id, state_id) when prestate_id === state_id, do: true
+
+  @doc """
   Approve a content by approval system
   """
-  require IEx
+
   @spec approve_content(User.t(), ApprovalSystem.t()) :: ApprovalSystem.t()
   def approve_content(
         current_user,
@@ -425,22 +513,24 @@ defmodule WraftDoc.Enterprise do
     Document.update_instance_state(current_user, instance, post_state)
 
     proceed_approval(approval_system)
-    |> Repo.preload([
-      :instance,
-      :pre_state,
-      :post_state,
-      :approver,
-      :user,
-      :organisation
-    ])
+    |> Repo.preload(
+      [
+        :instance,
+        :pre_state,
+        :post_state,
+        :approver,
+        :user,
+        :organisation
+      ],
+      force: true
+    )
   end
 
-  @doc """
-  Proceed approval make the status of approval system as approved
-  """
+  # Proceed approval make the status of approval system as approved
+
   @spec proceed_approval(ApprovalSystem.t()) :: ApprovalSystem.t()
-  def proceed_approval(approval_system) do
-    params = %{approved: true}
+  defp proceed_approval(approval_system) do
+    params = %{approved: true, approved_log: NaiveDateTime.local_now()}
 
     approval_system
     |> ApprovalSystem.approve_changeset(params)