+++ lib/wraft_doc_web/views/comment_view.ex
@@ -1,6 +1,9 @@
 defmodule WraftDocWeb.Api.V1.CommentView do
   use WraftDocWeb, :view
   alias __MODULE__
+  alias WraftDocWeb.Api.V1.ProfileView
+  alias WraftDocWeb.Api.V1.UserView
+  require IEx
 
   def render("comment.json", %{comment: comment}) do
     %{
@@ -10,6 +13,8 @@ defmodule WraftDocWeb.Api.V1.CommentView do
       parent_id: comment.parent_id,
       master: comment.master,
       master_id: comment.master_id,
+      user: render_one(comment.user, UserView, "user.json", as: :user),
+      profile: render_one(comment.user.profile, ProfileView, "base_profile.json", as: :profile),
       inserted_at: comment.inserted_at,
       updated_at: comment.updated_at
     }
@@ -30,6 +35,19 @@ defmodule WraftDocWeb.Api.V1.CommentView do
       page_number: page_number,
       total_pages: total_pages,
       total_entries: total_entries
+    }
+  end
+
+  def render("delete.json", %{comment: comment}) do
+    %{
+      id: comment.uuid,
+      comment: comment.comment,
+      is_parent: comment.is_parent,
+      parent_id: comment.parent_id,
+      master: comment.master,
+      master_id: comment.master_id,
+      inserted_at: comment.inserted_at,
+      updated_at: comment.updated_at
     }
   end
 end